Add files via upload
The To-Do List project is a web application developed to assist users in planning and managing their daily activities. Built using HTML, CSS, and JavaScript, this application allows users to create, add, and delete tasks efficiently. The project is part of the Brainwave Matrix Solutions Internship Program, designed to enhance productivity and task management skills.